refactor: extract BaselineEntry to separate file
Split baseline_manager.dart to follow one-class-per-file convention.
Sealed classes and their implementations remain grouped by design.
